requirements.txt
# Mr. Wlah (Write Like A Human) - Dependencies

# Flask for web server
Flask==2.3.3
Flask-Cors==4.0.0

# Environment variable management
python-dotenv==1.0.0

# Auth0 integration
authlib==1.2.1
python-jose==3.3.0

# MongoDB integration
pymongo==4.5.0
dnspython==2.4.2

# Google Gemini API
google-genai>=1.0.0

# File handling
PyPDF2==3.0.1
python-docx==0.8.11

# Utilities
requests==2.31.0
gunicorn==21.2.0 